# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AreaRule": {
        "additionalProperties": false,
        "properties": {
          "area_id": {
            "pattern": "^(syn|lon)-n[0-9a-z]+$",
            "title": "Area Id",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "rule": {
            "$ref": "#/components/schemas/AreaRuleKind"
          }
        },
        "required": [
          "area_id",
          "rule",
          "provenance"
        ],
        "title": "AreaRule",
        "type": "object"
      },
      "AreaRuleKind": {
        "enum": [
          "exclude",
          "only"
        ],
        "title": "AreaRuleKind",
        "type": "string"
      },
...
      "Budget": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "segment": {
            "$ref": "#/components/schemas/Segment"
          },
          "strictness": {
            "$ref": "#/components/schemas/Strictness"
          },
          "weight": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "amount",
          "segment",
          "strictness",
          "weight",
          "provenance"
        ],
        "title": "Budget",
        "type": "object"
      },
...
      "Combine": {
        "enum": [
          "slowest",
          "mean"
        ],
        "title": "Combine",
        "type": "string"
      },
      "Commute": {
        "additionalProperties": false,
        "properties": {
          "max_minutes": {
            "maximum": 120.0,
            "minimum": 10.0,
            "title": "Max Minutes",
            "type": "integer"
          },
          "mode": {
            "$ref": "#/components/schemas/Mode"
          },
          "place_id": {
            "pattern": "^(syn|lon)-p[0-9a-z]+$",
            "title": "Place Id",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "strictness": {
            "$ref": "#/components/schemas/Strictness"
          }
        },
        "required": [
          "place_id",
          "mode",
          "max_minutes",
          "strictness",
          "provenance"
        ],
        "title": "Commute",
        "type": "object"
      },
...
      "CompareBody": {
        "additionalProperties": false,
        "properties": {
          "area_ids": {
            "items": {
              "pattern": "^(syn|lon)-n[0-9a-z]+$",
              "type": "string"
            },
            "maxItems": 4,
            "minItems": 2,
            "title": "Area Ids",
            "type": "array"
          },
          "spec": {
            "$ref": "#/components/schemas/PreferenceSpec"
          }
        },
        "required": [
          "area_ids",
          "spec"
        ],
        "title": "CompareBody",
        "type": "object"
      },
...
      "CompareData": {
        "additionalProperties": false,
        "properties": {
          "areas": {
            "items": {
              "$ref": "#/components/schemas/ComparedArea"
            },
            "title": "Areas",
            "type": "array"
          },
          "facts": {
            "items": {
              "$ref": "#/components/schemas/Fact"
            },
            "title": "Facts",
            "type": "array"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/CompareRow"
            },
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "areas",
          "rows",
          "facts"
        ],
        "title": "CompareData",
        "type": "object"
      },
...
      "ComparedArea": {
        "additionalProperties": false,
        "properties": {
          "area_id": {
            "pattern": "^(syn|lon)-n[0-9a-z]+$",
            "title": "Area Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CompareStatus"
          }
        },
        "required": [
          "area_id",
          "name",
          "status"
        ],
        "title": "ComparedArea",
        "type": "object"
      },
...
      "Envelope_CompareData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CompareData"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "Envelope[CompareData]",
        "type": "object"
      },
      "Envelope_ExplanationsData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ExplanationsData"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "Envelope[ExplanationsData]",
        "type": "object"
      },
...
      "Envelope_RankData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RankData"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "Envelope[RankData]",
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldProblem"
            },
            "title": "Fields",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "fields"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "meta",
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ExplanationsBody": {
        "additionalProperties": false,
        "properties": {
          "limit": {
            "default": 3,
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "spec": {
            "$ref": "#/components/schemas/PreferenceSpec"
          }
        },
        "required": [
          "spec"
        ],
        "title": "ExplanationsBody",
        "type": "object"
      },
      "ExplanationsData": {
        "additionalProperties": false,
        "properties": {
          "explanations": {
            "items": {
              "$ref": "#/components/schemas/Explanation"
            },
            "title": "Explanations",
            "type": "array"
          },
          "facts": {
            "items": {
              "$ref": "#/components/schemas/Fact"
            },
            "title": "Facts",
            "type": "array"
          }
        },
        "required": [
          "explanations",
          "facts"
        ],
        "title": "ExplanationsData",
        "type": "object"
      },
...
      "FeatureWeight": {
        "additionalProperties": false,
        "properties": {
          "direction": {
            "$ref": "#/components/schemas/Direction"
          },
          "feature_id": {
            "$ref": "#/components/schemas/FeatureId"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "weight": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "feature_id",
          "weight",
          "direction",
          "provenance"
        ],
        "title": "FeatureWeight",
        "type": "object"
      },
...
      "Meta": {
        "additionalProperties": false,
        "properties": {
          "engine_version": {
            "title": "Engine Version",
            "type": "string"
          },
          "release_id": {
            "pattern": "^(syn|lon)-\\d{4}-\\d{2}-\\d{2}-\\d{2}$",
            "title": "Release Id",
            "type": "string"
          },
          "synthetic": {
            "title": "Synthetic",
            "type": "boolean"
          }
        },
        "required": [
          "release_id",
          "engine_version",
          "synthetic"
        ],
        "title": "Meta",
        "type": "object"
      },
...
      "Operations": {
        "additionalProperties": false,
        "properties": {
          "area_ops": {
            "items": {
              "$ref": "#/components/schemas/AreaEdit"
            },
            "title": "Area Ops",
            "type": "array"
          },
          "budget_ops": {
            "items": {
              "$ref": "#/components/schemas/BudgetEdit"
            },
            "title": "Budget Ops",
            "type": "array"
          },
          "commute_ops": {
            "items": {
              "$ref": "#/components/schemas/CommuteEdit"
            },
            "title": "Commute Ops",
            "type": "array"
          },
          "setting_ops": {
            "items": {
              "$ref": "#/components/schemas/SettingEdit"
            },
            "title": "Setting Ops",
            "type": "array"
          },
          "tag_ops": {
            "items": {
              "$ref": "#/components/schemas/TagEdit"
            },
            "title": "Tag Ops",
            "type": "array"
          },
          "weight_ops": {
            "items": {
              "$ref": "#/components/schemas/WeightEdit"
            },
            "title": "Weight Ops",
            "type": "array"
          }
        },
        "required": [
          "budget_ops",
          "commute_ops",
          "weight_ops",
          "tag_ops",
          "area_ops",
          "setting_ops"
        ],
        "title": "Operations",
        "type": "object"
      },
...
      "PreferenceSpec": {
        "additionalProperties": false,
        "properties": {
          "areas": {
            "items": {
              "$ref": "#/components/schemas/AreaRule"
            },
            "title": "Areas",
            "type": "array"
          },
          "budget": {
            "$ref": "#/components/schemas/Budget"
          },
          "commute_combine": {
            "$ref": "#/components/schemas/Combine"
          },
          "commute_combine_from": {
            "$ref": "#/components/schemas/Provenance"
          },
          "commute_weight": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Commute Weight",
            "type": "number"
          },
          "commute_weight_from": {
            "$ref": "#/components/schemas/Provenance"
          },
          "commutes": {
            "items": {
              "$ref": "#/components/schemas/Commute"
            },
            "maxItems": 3,
            "title": "Commutes",
            "type": "array"
          },
          "pt_basis": {
            "$ref": "#/components/schemas/PtBasis"
          },
          "pt_basis_from": {
            "$ref": "#/components/schemas/Provenance"
          },
          "schema_version": {
            "const": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagWeight"
            },
            "title": "Tags",
            "type": "array"
          },
          "tenure": {
            "$ref": "#/components/schemas/Tenure"
          },
          "tenure_from": {
            "$ref": "#/components/schemas/Provenance"
          },
          "weights": {
            "items": {
              "$ref": "#/components/schemas/FeatureWeight"
            },
            "title": "Weights",
            "type": "array"
          }
        },
        "required": [
          "schema_version",
          "tenure",
          "budget",
          "commutes",
          "commute_combine",
          "pt_basis",
          "commute_weight",
          "weights",
          "tags",
          "areas",
          "tenure_from",
          "commute_combine_from",
          "pt_basis_from",
          "commute_weight_from"
        ],
        "title": "PreferenceSpec",
        "type": "object"
      },
...
      "Provenance": {
        "enum": [
          "stated",
          "inferred",
          "default",
          "ui_edit"
        ],
        "title": "Provenance",
        "type": "string"
      },
      "PtBasis": {
        "enum": [
          "typical",
          "just_missed"
        ],
        "title": "PtBasis",
        "type": "string"
      },
      "RankBody": {
        "additionalProperties": false,
        "properties": {
          "limit": {
            "default": 20,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "operations": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Operations"
              },
              {
                "type": "null"
              }
            ]
          },
          "spec": {
            "$ref": "#/components/schemas/PreferenceSpec"
          }
        },
        "required": [
          "spec"
        ],
        "title": "RankBody",
        "type": "object"
      },
      "RankData": {
        "additionalProperties": false,
        "properties": {
          "applied": {
            "items": {
              "$ref": "#/components/schemas/Applied"
            },
            "title": "Applied",
            "type": "array"
          },
          "empty_spec": {
            "title": "Empty Spec",
            "type": "boolean"
          },
          "filtered": {
            "items": {
              "$ref": "#/components/schemas/Filtered"
            },
            "title": "Filtered",
            "type": "array"
          },
          "ranked": {
            "items": {
              "$ref": "#/components/schemas/RankedArea"
            },
            "title": "Ranked",
            "type": "array"
          },
          "rejected": {
            "items": {
              "$ref": "#/components/schemas/Rejected"
            },
            "title": "Rejected",
            "type": "array"
          },
          "scores": {
            "items": {
              "$ref": "#/components/schemas/Score"
            },
            "title": "Scores",
            "type": "array"
          },
          "spec": {
            "$ref": "#/components/schemas/PreferenceSpec"
          },
          "spec_hash": {
            "title": "Spec Hash",
            "type": "string"
          },
          "unranked": {
            "items": {
              "$ref": "#/components/schemas/Unranked"
            },
            "title": "Unranked",
            "type": "array"
          }
        },
        "required": [
          "scores",
          "ranked",
          "filtered",
          "unranked",
          "empty_spec",
          "spec",
          "spec_hash",
          "applied",
          "rejected"
        ],
        "title": "RankData",
        "type": "object"
      },
...
      "TagWeight": {
        "additionalProperties": false,
        "properties": {
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "tag_id": {
            "$ref": "#/components/schemas/TagId"
          },
          "weight": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "tag_id",
          "weight",
          "provenance"
        ],
        "title": "TagWeight",
        "type": "object"
      },
...
      "Tenure": {
        "enum": [
          "rent",
          "buy"
        ],
        "title": "Tenure",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Ranks named neighbourhoods for a preference spec and shows its working. Every response says whether the data behind it is synthetic.",
    "title": "Burro API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/compare": {
      "post": {
        "description": "Two to four areas side by side, the rows in the order of the person's own weights.",
        "operationId": "compare",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_CompareData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Content Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Compare"
      }
    },
    "/v1/explanations": {
      "post": {
        "description": "A few checked sentences about each of the top areas, and the facts they cite.",
        "operationId": "explain_top",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplanationsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ExplanationsData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Content Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Explain Top"
      }
    },
...
    "/v1/rank": {
      "post": {
        "description": "Apply any edits to the spec, then rank every area of the release for it.",
        "operationId": "rank",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RankBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RankData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Content Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Rank"
      }
    },
...
  }
}